Rewrite the scrollbar rendering system.
Changes include:

There are only two faces, Scrollbar and ScrollThumb. The background of
ScrollThumb is used for the region that represents the extent of the current
view, the background of Scrollbar is used for everything else. The foreground of
each face is used to display an icon representing the number of selections in
that portion of the document.

There's only one option, selection_chars, which is a space-separated list of
characters that will be used as icons to represent the number of selections in a
region of the scrollbar. For example, if the option is set to "0 1 2 M", then
"0" will be displayed in portions of the scrollbar with no selections, "1" in
portions with one selection, "2" in portions with two selections, and "M" in
portions with three or more. If the option begins with a space, an actual space
will be used to represent "no selections".